+++ tosshin_cpp/include/tosshin_cpp/provider/navigation_provider.hpp
@@ -18,8 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef TOSSHIN_CPP__NAVIGATION__NAVIGATION_PROVIDER_HPP_
-#define TOSSHIN_CPP__NAVIGATION__NAVIGATION_PROVIDER_HPP_
+#ifndef TOSSHIN_CPP__PROVIDER__NAVIGATION_PROVIDER_HPP_
+#define TOSSHIN_CPP__PROVIDER__NAVIGATION_PROVIDER_HPP_
 
 #include <rclcpp/rclcpp.hpp>
 #include <tosshin_interfaces/tosshin_interfaces.hpp>
@@ -173,4 +173,4 @@ const Maneuver & NavigationProvider::on_configure_maneuver(const Maneuver & mane
 
 }  // namespace tosshin_cpp
 
-#endif  // TOSSHIN_CPP__NAVIGATION__NAVIGATION_PROVIDER_HPP_
+#endif  // TOSSHIN_CPP__PROVIDER__NAVIGATION_PROVIDER_HPP_
